Refactor Replica state
This moves Replica state types to the state.rs file, and changes
`Replica` so that it always includes a `LeaderState`, `CandidateState`,
and `FollowerState` struct. The reason to include the respective states
regardless of the current mode of the `Replica` is to avoid having to
create a new state instance every time the `Replica` changes mode.