introduce message wrapper, cleanup client
This commit introduces a wrapper message type, `Message`, which all
messages sent between Raft clients and servers should use to wrap the
message type.  Additionally, the client has been reduced to just the
append method. Leader refresh was removed because it's impossible to
provide correctly, and die was removed because its uneeded right now,
and in the long run shutdown should probably be provided by the server
objects, not the rpc system.