properly replace peer connections
Server makes sure that we always have exactly one Connection object
assigned to a peer at a time.  The connection may be unconnected if
it has recently been reset, in which case a reconnection timeout will be
registered.

This commit makes sure that when a connection is received from another
peer, that the existing connection is removed.  This is a bit
simplistic, and won't work once we have configuration change (peers
attempting to join the cluster), but for now it is good enough.