Reduce test flakiness on OSX
Many of the server.rs tests will fail eventually if run in a loop on OSX. This
commit reduces the likelyhood of failures, but issues still remain. The
problem manifests as failed calls to `stream_shutdown` or `client_connected`.
The test may also hang indefinitely, which is a side effect of how we are
currently testing for shutdown in `stream_shutdown`.

It's unclear at this point whether this is due to misuse of the MIO APIs in
server.rs or because of a bug in MIO, or even kqueue itself.

Potentially incomplete list of tests which remain flaky after this patch:

* `test_invalid_client_message`
* `test_invalid_accept`
* `test_unreachable_peer_reconnect`
* `test_peer_accept`