Tweak logging messages; remove scopes from connection.rs
The connection.rs logging scopes were hard to follow, and since each
call method on Connection is typically a leaf-call, scopes are not
necessary. For now Connection scopes are removed, if the call graph on
Connection methods gets more complicated in the future the scopes may
need to be added back.