Leader: increment peer next_index after queuing client proposal
In order to support pipelining (multiple in-flight append entry requests from
the leader to followers), we set the `next_index` of each follower immediately
upon queueing a message to send to that follower, not upon receiving
confirmation that the message was received. This commit adds a missing
`next_index` increment during client proposal.